refactor: extract helpers and ipfs handlers from writers.ts
